Module type Block.BlockRewards

Module that contains the reward function to be called when creating a new block.

val reward : int -> balances -> balances

Given the minter of the block and the balances, and produce the resulting balances.

parameter minter
the node creating the block

parameter balances
the current balances of each node