Module Statistics.Make

module type Arg = sig ... end

module Aux : sig ... end

module Average (X : Arg) : Stats with type t = int

Create an implementation for a Statistics module that computes an Average.

module Median (X : Arg) : Stats with type t = int

Create an implementation for a Statistics module that computes a Median.

module Percentile95 (X : Arg) : Stats with type t = int

Create an implementation for a Statistics module that computes a 95th Percentile.

module Max (X : Arg) : Stats with type t = int

Create an implementation for a Statistics module that computes a Maximum.

module Min (X : Arg) : Stats with type t = int

Create an implementation for a Statistics module that computes a Minimum.

module CountPerNode (X : Arg) : Stats with type t = int

Create an implementation for a Statistics module that computes a Counter per Node.

module CountAll (X : Arg) : Stats with type t = int

Create an implementation for a Statistics module that computes a Global Counter for integers.

module CountAllF (X : Arg) : Stats with type t = float

Create an implementation for a Statistics module that computes a Global Counter for floats.