Module Unique.Make

Create an implementation for Unique, given a wrapper T for a type of values.

Parameters

module T : T

Signature

type v = T.v

Type of the values being converted to unique integers.

val id : v -> int

Get an integer that uniquely identifies v.

parameter x
the value for which we want a unique identifier