Module type Unique.Unique

Signature of the unique module. Purpose is to get a unique identifier for any value of type v.

type v

Type of the values being converted to unique integers.

val id : v -> int

Get an integer that uniquely identifies v.

parameter x
the value for which we want a unique identifier